// packages/ui/src/styles/theme.d.ts
/** Brand colors used for theming. */
export interface BrandTheme {
    backgroundColor: {
        base: string;
        card: string;
        invert: string;
        destructive: string;
    };
    textColor: {
        primary: string;
        secondary: string;
        accent: string;
        invert: string;
        destructive: string;
    };
    borderColor: {
        default: string;
    };
}
export declare const DEFAULT_LIGHT_THEME: BrandTheme;
export declare const DEFAULT_DARK_THEME: BrandTheme;